REFACTOR 1.1
